backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-multipart==0.0.24
python-dotenv==1.2.2
PyMuPDF==1.27.2.2
python-docx==1.2.0
spacy==3.8.14
rapidfuzz==3.14.5
scikit-learn==1.7.2
google-genai==1.72.0
requests==2.32.3
beautifulsoup4==4.12.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl